--- src/NoteEditor/BlockRenderer/Block.css
@@ -1,96 +1,0 @@
-.blocks {
-  width: 100%;
-}
-.block {
-  position: relative;
-  outline: none;
-  width: 100%;
-}
-.block:not(.block-text):focus-visible:after,
-.block:not(.block-text):focus:after {
-  content: '';
-  position: absolute;
-  width: calc(100% - 4px);
-  height: calc(100% - 4px);
-  left: 0;
-  top: 0;
-  border-radius: 5px;
-  border: 2px solid #646cff;
-  pointer-events: none;
-}
-.block.block-text:focus {
-  background-color: #eee;
-}
-.block.block-text {
-  padding: 12px;
-  width: calc(100% - 24px);
-  white-space: pre-wrap;
-  overflow-wrap: anywhere;
-}
-/* Text formatting */
-.block.block-text .bold {
-  font-weight: bold;
-}
-.block.block-text .italic {
-  font-style: italic;
-}
-.block.block-text .underline {
-  text-decoration: underline;
-}
-.block.block-text .strikethrough {
-  text-decoration: line-through;
-}
-.block.block-text .underline.strikethrough {
-  text-decoration: underline line-through;
-}
-.block.block-text .code {
-  font-family: monospace;
-  background-color: rgba(27, 31, 35, 0.05);
-  padding: 0.2em 0.4em;
-  border-radius: 6px;
-}
-.block.block-text.initial::before {
-  content: 'Add text, drawings, files...';
-  color: #555;
-}
-
-.block.block-text.initial:focus::before {
-  content: '';
-}
-
-.block.block-media {
-  width: 100%;
-  overflow: hidden;
-  display: flex;
-  justify-content: center;
-  margin: 12px 0;
-}
-.block.block-media img {
-  width: 100%;
-  height: auto;
-  max-width: 50vh;
-  margin: 0 auto;
-  border-radius: 5px;
-  animation: popIn 0.5s cubic-bezier(0.075, 0.82, 0.165, 1);
-}
-
-@keyframes popIn {
-  0% {
-    transform: translateY(500px);
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-.debug-preview {
-  width: calc(100% - 2em);
-  overflow-x: hidden;
-  font-family: monospace;
-  white-space: pre-wrap;
-  padding: 1em;
-  margin: 0;
-  width: 500px;
-  overflow-wrap: anywhere;
-}
