--- src/NoteEditor/BlockRenderer/blocks/ScribbleBlock.tsx
@@ -1,13 +1,0 @@
-import Note from '../../../lib/Note';
-import type { ScribbleBlock as ScribbleBlockType } from '../../../lib/types';
-import '../Block.css';
-
-function ScribbleBlock({ note, block }: { note: Note; block: ScribbleBlockType }) {
-  return (
-    <div className="block block-scribble">
-      <span>TODO: scribble</span>
-    </div>
-  );
-}
-
-export default ScribbleBlock;
