+++ src/NoteEditor/Editor/Toolbar/Toolbar.css
@@ -37,6 +37,16 @@ hr {
   gap: 2px;
   background-color: #ccc;
   border-radius: 50px;
+  height: 100%;
+}
+.toolbar .tools {
+  height: calc(100% - 8px);
+}
+.toolbar .tool-label {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0 5px;
 }
 button.format {
   border-radius: 6px;
